requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cfgv==3.4.0
charset-normalizer==3.4.3
coverage==7.10.6
distlib==0.4.0
Django==5.2.5
django-prometheus==2.4.1
django-restframework==0.0.1
djangorestframework==3.16.1
filelock==3.19.1
identify==2.6.14
idna==3.10
iniconfig==2.1.0
nodeenv==1.9.1
packaging==25.0
platformdirs==4.4.0
pluggy==1.6.0
pre_commit==4.3.0
prometheus_client==0.22.1
Pygments==2.19.2
pytest==8.4.2
pytest-cov==6.3.0
pytest-django==4.11.1
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.5
ruff==0.12.12
sqlparse==0.5.3
urllib3==2.5.0
virtualenv==20.34.0
